Updated rand from 0.3 to 0.6 (#5)
* update rand to v0.6

* rand v0.6 adjustments

+++ src/cli.rs
@@ -61,7 +61,7 @@ pub struct Expression {
     threads: usize,
     cells: Regex,
     window_width: u32,
-    seed: usize,
+    seed: u64,
     #[builder(setter(skip))]
     counts: collections::HashMap<String, collections::HashMap<String, Counts>>,
 }
